requirements.txt
absl-py==1.4.0
anyio==3.6.2
arandr==0.1.11
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
autocommand==2.2.2
Automat==22.10.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
btrfsutil==6.2.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.0
colorama==0.4.6
comm==0.1.2
ConfigSpace==0.6.0
constantly==15.1.0
contourpy==1.0.7
cryptography==39.0.2
cvlib==0.2.7
cycler==0.11.0
Cython==0.29.33
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
deluge==2.1.1
deprecation==2.1.0
distro==1.8.0
emcee==3.1.4
entrypoints==0.4
executing==1.2.0
facenet-pytorch==2.5.2
fastjsonschema==2.16.3
filelock==3.10.0
flatbuffers==23.3.3
fonttools==4.39.0
gast==0.4.0
gdown==4.6.4
GeoIP==1.3.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
hyperlink==21.0.0
idna==3.4
ifaddr==0.2.0
imageio==2.26.0
imutils==0.5.4
incremental==22.10.0
inflect==6.0.2
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.text==3.11.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.13
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.3.0
jupyter_packaging==0.12.3
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.20.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
libtorrent==2.0.8
lit==15.0.7
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
more-itertools==9.0.0
mpmath==1.3.0
mtcnn==0.1.1
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
netsnmp-python==1.0a1
networkx==3.0
nftables==0.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openbox==0.8.1
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.0
Platypus-Opt==1.0.4
ply==3.11
prettytable==3.6.0
progressbar==2.5
prometheus-client==0.15.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pwquality==1.4.5
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.23.0
pycparser==2.21
pydantic==1.10.6
Pygments==2.14.0
PyGObject==3.42.2
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.11.1
PyQtWebEngine==5.15.6
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
pyxdg==0.28
PyYAML==5.4
pyzmq==24.0.1
qutebrowser==2.5.3
ranger-fm==1.9.3
Reflector==2021.11.20.2.41.3
rencode==1.0.6
requests==2.28.2
requests-oauthlib==1.3.1
requests-unixsocket==0.3.0
retina-face==0.0.13
rich==13.3.2
rsa==4.9
scikit-learn==1.2.2
scikit-optimize==0.9.0
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
service-identity==21.1.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
statsmodels==0.13.5
sympy==1.11.1
team==1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
testpath==0.6.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
trove-classifiers==2023.3.10
Twisted==22.10.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.15
validate-pyproject==0.12.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.15.0
zope.interface==5.5.2
